requirements.txt
docker-py>=1.6.0
ipaddress==1.0.16
requests>=2.5.3
python-dateutil>=2.4.0
six>=1.9.0
pyyaml==3.10
argparse==1.3.0
psutil==3.4.2
